+++ core/src/main/resources/org/akaza/openclinica/applicationContext-core-service.xml
@@ -38,6 +38,10 @@
         <constructor-arg ref="dataSource"/>
     </bean>
 
+    <bean id="participateServiceImpl" class="org.akaza.openclinica.service.ParticipateServiceImpl">
+        <constructor-arg ref="dataSource"/>
+    </bean>
+
     <bean id="importCRFDataService" class="org.akaza.openclinica.web.crfdata.ImportCRFDataService">
         <constructor-arg ref="dataSource"/>
     </bean>
